Surface separate lambda and ecs cluster subnet lists
Internally, the Cumulus Terraform module uses separate variables for
subnet IDs for Lambdas and subnet IDs for the ECS cluster, but this
example deployment was using only 1 list of subnet IDs for both
purposes.  One case where this can be limiting is when a user wants
`private_archive_api_gateway` is set to `false` and also specify an
empty list of subnet IDs for Lambdas while specifying at least one
subnet ID for the ECS cluster, which is required.